\name{dynOmics-package}
\Rdversion{1.14}
\alias{dynOmics-package}
\alias{dynOmics}
\docType{package}
\title{
Fast Fourier transform to estimate delays in expression initiation to identify associations between time course 'omics' data.
}
\description{
The package provides functions to identify associations within one or between two time course 'omics' data and visualise the associations : associateData to estimate the delays and identify associations of data sets containing time course 'omics' experiments; plot.associations: to visualise associated profiles.  
}
\details{
\tabular{ll}{
Package: \tab dynOmics\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2018-06-12\cr
License: \tab GPL-2 \cr
LazyLoad: \tab yes\cr
}
Functions for associating data: \code{\link[dynOmics]{associateData}} \cr
Functions for summarization: \code{\link[dynOmics]{summary.associations}}\cr
Functions for plots: \code{\link[dynOmics]{plot.associations}}\cr
}
\author{
Jasmin Straube with contributions from Kim-Anh Le Cao, Emma Huang and Anne Bernard

Maintainer: Jasmin Straube <j.straube@qimrberghofer.edu.au>
}
\keyword{package}
