% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.associations-methods.R
\name{summary.associations}
\alias{summary.associations}
\title{Summary of a \code{associations} Object}
\usage{
\method{summary}{associations}(object, ...)
}
\arguments{
\item{object}{An object of class \code{associations} .}

\item{...}{Additional arguments which are passed to \code{summary}.}
}
\value{
summary of the \code{associations} object.
}
\description{
Summarises the \code{associations} object returned by the \code{\link{associateData}} method.
}
\examples{
\dontrun{
data(Metabolites)
data(Transcripts)
associations <- associateData(Metabolites[,1],Transcripts[,c(1:50)])
summary(associations)
}
}
